fix: 문제 및 테스트 코드 수정
