+++ TIL/daily/2025-03-17.md
@@ -43,7 +43,8 @@ class Hero {
 	required this._hp 불가능 -> named parameter는 밑줄로 시작할 수 없다
 	Hero(this._hp, {required this.name}) 가능
 	인스턴스 생성시 Hero(100, name: '홍길동')으로 생성
-	/*
+	*/
+}
 ```
 
 ---
